Update README.md
# Project Name

## Description

This project is centered around the implementation of the Technological Fields Theory (TFT) and the TFT Problem Solving Parameters (TFTpsp). The TFTpsp are 33 parameters used for solving problems and accelerating technological and scientific progress for improving people's lives and the natural environment. The parameters are incorporated into the structure of an artificial intelligence system to optimize its problem-solving abilities.

The key components of the project include:

- **Mapping of Technological Context:** This involves identifying existing and potential technologies that can be used or integrated to solve the problem.

- **Mapping of Problem Sources and Cause-Effect Relationships:** This involves analyzing the causes and effects of the problem, identifying relevant variables and their interactions.

- **Evaluation of Social and Economic Impacts of Proposed Solutions:** This involves estimating the benefits and costs of solutions for society and the economy, taking into account ethical, legal, and environmental aspects.

- **Defining the Implementation and Dissemination of Innovative Solutions:** This includes planning the modalities and strategies for realizing and spreading the solutions, involving interested stakeholders.

## Implementation

This project utilizes Python for implementing the above functionalities, and the Python files included in this repository are organized as follows:

- `holos.tft`: Contains functions and classes for managing TFTpsp, such as TFT-33, TFT-12, and TFT-31.

- `holos.cfu`: Contains functions and classes for managing the Universal Fundamental Code (CFU), a binary code that represents the laws of nature and the universe.

- `holos.dna`: Contains functions and classes for managing digital DNA, a digital genetic-epigenetic structure that regulates the functions of the digital organism using AI techniques.

- `holos.brain`: Contains functions and classes for managing the digital brain, a network of interconnected websites like neurons that transmit and process information.

- `holos.body`: Contains functions and classes for managing the physical body, based on the Internet of Things (IoT), which allows effective communication with IoT devices.

## Data Loading

Data, problems, and feedback are loaded using standardized formats or universal communication protocols such as MQTT or CoAP, but they can also be customized based on individual needs. For example, JSON or XML can be used to load data, MQTT or CoAP for problems, and CFU format for feedback.